Refactors distribution code and adds unit tests
