+++ jest.config.js
@@ -1,25 +1,5 @@
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
-  collectCoverageFrom: ["**/*.{js,jsx,ts,tsx}", "!**/*.d.ts", "!**/node_modules/**"],
-  setupFilesAfterEnv: ["<rootDir>/setupTests.js"],
-  testPathIgnorePatterns: ["/node_modules/", "/.next/"],
-  transform: {
-    "^.+\\.(js|jsx|ts|tsx)$": "<rootDir>/node_modules/babel-jest",
-    "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-  },
-  transformIgnorePatterns: ["/node_modules/", "^.+\\.module\\.(css|sass|scss)$"],
-  moduleNameMapper: {
-    "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy",
-  },
-  //additional prop from a github issue: https://github.com/zeit/next.js/issues/8663
-  preset: "ts-jest",
-  modulePathIgnorePatterns: ["studio", "public", "docs", ".next", "config", "coverage"],
-  testEnvironment: "jsdom",
-  globals: {
-    // we must specify a custom tsconfig for tests because we need the typescript transform
-    // to transform jsx into js rather than leaving it jsx such as the next build requires.  you
-    // can see this setting in tsconfig.jest.json -> "jsx": "react"
-    "ts-jest": {
-      tsConfig: "tsconfig.jest.json",
-    },
-  },
+  preset: 'ts-jest',
+  testEnvironment: 'node',
 };
