+++ package.json
@@ -12,7 +12,7 @@
     "coverage": "jest --coverage",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "test": "jest --transformIgnorePatterns \"node_modules/(?!@toolz/allow)/\" --env=jsdom --watch"
+    "test": "jest"
   },
   "dependencies": {
     "@auth0/auth0-react": "^1.9.0",
@@ -64,7 +64,7 @@
     "@types/enzyme": "^3.10.11",
     "@types/enzyme-adapter-react-16": "^1.0.6",
     "@types/istanbul-lib-coverage": "^2.0.4",
-    "@types/jest": "^27.4.0",
+    "@types/jest": "^27.5.2",
     "@types/luxon": "^2.0.9",
     "@types/node": "17.0.12",
     "@types/react": "17.0.38",
@@ -81,8 +81,8 @@
     "jest": "^27.5.1",
     "prettier": "2.6.2",
     "redux-devtools-extension": "^2.13.9",
-    "ts-jest": "^27.1.3",
-    "typescript": "4.5.5",
+    "ts-jest": "^27.1.5",
+    "typescript": "^4.5.5",
     "typescript-fsa": "^3.0.0",
     "webpack": "^5.70.0"
   }
